Move multiplayer handling to universe.js
